Salvage buff
Finding salvage for making guns is pain, this fixes it so that people don't have to do the funny glitch (which was fixed) if they want to make more than one gun, without spending the whole round doing so. Does NOT re-add tool salvage to cars, before people start whining about lag.